The sexpr evaluator must move, too
