+++ opencog/persist/sexpr/CMakeLists.txt
@@ -3,7 +3,6 @@
 ADD_LIBRARY (sexpr
 	AtomSexpr.cc
 	FrameSexpr.cc
-	SexprEval.cc
 	ValueSexpr.cc
 )
 
@@ -11,7 +10,6 @@ ADD_DEPENDENCIES(sexpr opencog_atom_types)
 
 TARGET_LINK_LIBRARIES(sexpr
 	atomspace
-	execution
 	atombase
 	${COGUTIL_LIBRARY}
 )
@@ -22,7 +20,6 @@ INSTALL (TARGETS sexpr EXPORT AtomSpaceTargets
 
 INSTALL (FILES
 	Sexpr.h
-	SexprEval.h
 	DESTINATION "include/opencog/persist/sexpr"
 )
